Add a function to repack a tarball or dir to a .tar.gz.
When an import has been done the .orig.tar.gz is needed. This function is the
first step towards that as it can do the repacking of most types. Currently
zip is unhandled, as zipfile is horrible, but it will be handled eventually.

Also add suoport for plain .tar archives. This now works for both
merge_upstream and repack_tarball.

Throw more specific errors for missing or present files.