Start using tags to mark the upstream imports.
Asking the user for the revision of the last upstream import isn't pretty,
so start marking them with tags. This isn't perfect, as it requires
--dirstate-tags, but that has been available for a while.

The UI needs to be tidied up, with more information when things go wrong,
and a way to override it if there was no import of the last upstream or
similar.