Extend the import_dsc code to handle .dsc with no .orig.tar.gz.
For the .dsc files with no .orig.tar.gz the upstream import must be looked up.
The patch must be applied to the parent tree, but then use the last revision
as a parent.

Switch to calling out to patch and lsdiff to do the patching and give the
information about what was touched.