Support importing the second set of .dsc.
Import a second set of .dsc files, with a new upstream. The new upstream
has the last upstream version as a parent, and the associated diff is added
on top, and gets a second parent of the previous head. This matches the
behaviour of merge_upstream.

Also fix the sorting of .dsc files.