Add an (untested) import-snapshot command that imports from snapshot.d.n.
The command gets the old versions of the package from s.d.n and imports them
all in to a new branch for you.

Also fix up the import code to only import a file once.

There needs to be some testing for this, but I need to work out how to do it,
and there will probably be some complexity in setting it up. Also the way
the implementation is done means that there is lots of special cases. These
should be removed.