Use uscan to retrieve upstream tarballs.
If a tarball is not found in the orig-dir and the user has a watch file then
use uscan to grab the tarball from upstream, and then repack it if needed
and place it in the orig-dir so that the build can continue.

If you are in merge mode then a dch -v will be all you need to try and build
a new upstream version.

This totally ignores signature checking and the like, but I am not sure how
I could handle that.