Add tests for non-native -> native transition on import.
When transitioning from a non-native package to native the first native version
is a merge of the last upstream and the last packaging changes. This required
pulling the import_tar code out of bzrtools, so that when a file moves from
the packaging branch to the native branch the file id is maintained.

I am not convinced that all cases are handled here, but it is a start.