Make the config say 'native = True' when it is so that the packages will build.
This does remove the config file when it transitions to non-native, so it
might play funny buggers with merges a bit, but it's a rare situation.

Also add some helpers in the tests for common things. Doesn't reduce the line
count much, but it makes the intention a bit clearer for some.