Internal hook execution support.
Hooks are run from the [HOOKS] section of the config files, with the normal
precedence rules.

Undefined hooks are ignored, and failing hooks raise an exception.

There is no policy defined here on the names of hooks, nor where they are
executed, and passed hook name will be looked up, and the working directory
for the hook can also be passed.