Hook up the hooks to the builddeb command.
There are currently 3 hooks, 'pre-export' that is run before the the package
is exported so that the branch/working tree can be modified. However the
tree is grabbed first if -r is specified, which may be suprsing. It is run
with the branch as the cwd.

The next two are 'pre-build' and 'post-build' that are run before and after
the builder, with the exported source as the working directory.

More points could be added, and the other commands may want to hook in to the
hooks, especially bd-do.

There could also be an option to disable hooks, for one run on the command
line, or through the config files, and perhaps an option to enable them, as
discussed in the spec.