Make the test suite pass again for bzr 0.92.
* test_webdav.py: 
Fix import statements, except for HttpServer which is both a
module and a class inported in bzrlib.tests which make it
impossible to import as a module.
(TestingDAVRequestHandler.send_head): Override send_head because
python-2.5 version implements the silly behavior of apache
redirecting directories when they do not end with a slash.

* __init__.py: 
Fix import statements.