Remove the -dev from the python build dependency.
We don't need the headers, just all versions, so python-all is fine.
However this introduces two lintian warnings. The first is that
one of python | python-dev | python-all-dev should be in build-depends.
The first could be added, but I didn't.

The second is that python-all is not needed in Build-Depends, just
Build-Depends-Indep, as no architecture specific packages are built. However
it is used to run the clean target, so we need it. This is a lintian false
positive.