Update how 'bzr missing' works when given --mine-only or --theirs-only
It used to always determine missing revisions on both sides, and then just suppress
displaying them if the flag was given. It would also return status 1 if either side
had extra revisions. However, if you are asking for --mine-only, it doesn't make sense
to return status 1 when the local branch has nothing new.
And the new code doesn't know whether there are new revisions on the other side anyway.