Classify bzr-fast-export
This file was getting a bit messy with all the global variables and parameter
passing.

This commit changes bzr-fast-export to be a real class. This simplifies
various methods and will allow easier hacking in the future.

There is a single change, though: the next_available_branch_name has been
simplified to just return a temporary branch name, as they aren't used for
anything else.