bzr-fast-export.py: Enhance marks support and add --checkpoint
This introduces a --marks= option that will set both the --export-marks and
the --import-marks. The program will just warn when --import-marks is set but
no marks file can be found.

Furthermore, this introduces a --checkpoint option. By default set at N=1000,
it will send a "checkpoint" every N revisions, and will also export its own
marks.

Using this, it is possible to resume an export if it has crashed before.