bzr-fast-export.py: Add support for ghost commits
This adds support for ghost commits. It does this by removing any parent that
is not currently in the repository. Thus, you cannot figure out
post-conversion that the commit was once a merge.

But hey, at least it doesn't crash :)

This also simplifies emit_commit somewhat, making it easier to follow.
However, this change does make emit_commit recursive, which might introduce
problems on large commits.