bzr-fast-export.py: Skip over broken commits.
This is a bit of a hack. The Emacs repository on http://bzr.notengoamigos.org
has a broken commit at revision 50255. This change will still output that
revision, but without any changes. The next revisions will skip it and diff
against its parent, thereby still resulting in a good conversion.