Add importing and exporting of marks to bzr-fastimport
This extends the GenericProcessor to be able to import and export revision
marks. This is useful when you want to do an incremental import of another
repository without rerunning the whole import all the time.

The file format used for the marks is the same as in bzr-fast-export. By
sharing each other's marks files, it is possible to set up a bidirectional
gateway with another repository.