Use file ids from other trees.
When there are parallel imports of files the file ids should be shared.

This change picks up the file ids from the other trees of the same type
(upstream or debian), and uses them if the file can't be found in any
parent.

It should perhaps be extended to just use a file id from any branch,
including one of a different type.