Use dpkg-source to extract the source packages and then import from that.
This should be much more robust, and means that we can import both upstream
and debianised parts using the same code path.

This means the TreePatcher can die, and we no longer have the ugly and
fragile hack to read the changelog from the .diff.gz.

The only downside is that we would now want a dpkg-source to be available
that was backed by the bzr VFS so that we can avoid having to bring files
to disk, but this isn't a great loss.