Only access the branch of other DistributionBranch, not the tree.
This allows them to be remote, or have no tree.

To do this we make them have .branch as primary object, not .tree,
and then never access .tree from another DistributionBranch.

.tree is then optional, and required if you actually want to do
any importing to the DistributionBranch.

There may be a better way to do the errors and enforce this, perhaps
through subclassing.