Read no more then 64k at a time in the smart protocol code.
The logic for this has been moved entirely into bzrlib.smart.medium, and
duplication (both in that module, and in bzrlib.smart.protocol) has been mostly
refactored out.  In particular there's now a SmartMedium base class used for
both client- and server-side media, and only one place that reading a line is
implemented.