Handle invalid version numbers in the changelog by ignoring them.
Any errors from parsing version numbers stops the parsing at that
point, so older versions are not considered as well.

This shouldn't be a problem with new packages, as new versions
shouldn't use an invalid version number.

It shouldn't be a problem for old revisions, as we probably won't
have history from before that point.