Pass --only-source to "apt-get source" call
This will prevent apt from assuming the package name we pass is a binary
package name which it must look up the source for. This would cause problems
when there are a source package and binary package that are named the same,
but the binary package isn't built from the source package, e.g. firefox,
linux. (bug #275446)