bzr-fast-export: Use an array for rename bookkeeping
We used to use a dictionary to keep track of recursive renames.
This bugs out because dictionaries in Python can have an arbitrary order
(at least they are not ordered on insertion time). This means that some
renames might be checked for before their parent rename was emitted, in which
case the rename would not be detected.

We fix this by using an array, which should always check for the renames in
the correct order.