Also merge tags when initialising one upstream from another.
If we don't merge the tags here then they are only visible in older
branches.

Also it can cause a failure if the changelog drops a later upstream
version from it, as the importer goes back to an older upstream version
to work from , and the tag has been dropped.