Add logic to map() so that it can also collapse when necessary.
It would be nice if check_remap could use size information to determine this
without actually having to build up a LeafNode, but that will depend on the
serializer in the end. Profiling will show if it ever really matters.

One other possibility would be to limit the number of new nodes we pull in at
a time. If we are going to have to load 50 LeafNodes then we are rather positive
that it won't fit into a new LeafNode.