Add LRUCache.resize(), and change the init arguments for LRUCache.
The old name was a bit confusing, and caused LRUSizeCache to re-use variables in
a confusing way with LRUCache.


Also, this changes the default cleanup size to be 80% of max_size. This should
be better, as it means we get a little bit of room when adding keys,
rather than having to cleanup after every add, we can instead do it in
batches.