Don't use the distribution name in the tags, fixing myriad issues.
This does give us a compatibility headache though. Old clients won't see
the new tags, but using the old tags lead to bugs, and the only way to
fix that is to use the new tags. Not a good situation, but moving forwards
we will have a much better situation.