Kill off export-upstream mode.
merge-upstream will be improved to merge the upstream branch, store it with
pristine-tar, and then produce the tarball as needed.

Using export-upstream for snapshots is now incorrect. You should use
"merge-upstream" every time instead, which will track if there is nothing
to do, and handle conflicts. You can autocommit the merge and build the
result if you like.