Change the code a little bit.
If a given text has not been seen before, insert all lines for that text.
At present, we are doing *worse* than knit compression, because we have
so many matching groups from various locations. Which causes us to
just have huge swaths of copies.

By inserting the full lines, we get more regions that we are able to
generate a larger match against.

This slows down the processing (10m => 24m), but improves compression
(16MB => 12MB).