fix up the failing tests.
The new delta code needs a 16-byte window to match, so to *know* that there will
be a match, you need ~32-bytes in common. (guarantees that 16-bytes somewhere in
that 32-byte range will match.)
Also, when setting 'max_delta', it is possible that we run out of bytes before
we actually find the last match, which would make things compress better.
This is rare in practice, because texts are longer than 40 bytes. But it happens
in testing.