Fix the root cause of why _iter_nodes was not handling key_filter.
The _search_prefix_filter was not evaluating _search_key_func(key) when
the key was not the same length as the internal key width.
Now we cap the max prefix at the width of a key, but if you are
searching for just a prefix portion of the overall key, you don't want the
extra nulls that _search_key() gives.
Added some chk_map tests to prove this.