Change the GroupCompressBlock code to allow not recording 'end'.
As long as we know what gc block we need to read, we don't care about how long
a record is. As it is already recorded in the data stream.
We already have read the whole compressed block into memory, so we aren't saving
a round trip. Though if we want trivial stripping of unused content,
we may want to keep the extra data in the index.
It still is good to have the functionality available to the extract() function.