Another disk format change.
We have pretty much settled on no-labels, at least until we implement
a gc+chk index (and then we can decide whether we want a fatter index
or to have more content in the group).

As such, we don't need to track the size of the header in the disk
bytestream anymore.

I decided to stick with ascii lengths in the outer wrapper.
Even large repositories have a small number of groups, so there isn't
many bytes that are actually wasted at that level. And it means
if you 'vim foo.pack' there are bits that can actually be interpreted.