Use a hybrid pass to walk all nodes.
This avoids a heapq, which we didn't really need anyway.
It is ~ the same number of passes, only we trade off a simplistic
'find_tails' call, with doing more work on the 1-pass over the dict.

It seems to be slightly faster for bzr, and slightly slower for OOo.
Not particularly noticeable in the noise of heads(), but measurable for
__init__ performance.