We don't need self._tails anymore, nor does it need to be a set.
Bring back '_find_tails' because walking the dictionary is faster than pushing/popping
into the set. Also we can use a list rather than a set, and populate it right away.
Down to 42.8ms bzr.dev and 550ms for OOo.
Note that _initialize_nodes time is 35ms and 407ms, and that _find_tails is approx 40ms for OOo.
So we are down to ~8ms for _find_tails + _find_gdfo for bzr.dev, which is probably
about as low as we can make it.
Any further improvements are going to be from _initialize_nodes.