Finish adding CHKInventory as a permutation in per_inventory.
Get the Read tests working. Change the apply_delta tests to be create_by_apply_delta
and implement that functionality on the regular Inventory class.
(just does a .copy() and then apply delta on the result.)

This also required fixing a lot of the tests that never asserted the objects
were valid.
Namely: InventoryFile('file-id', 'name', 'parent-id')
does not generate a fully valid file object, as it has no text_size or text_sha1.


The filter() tests are finally failing, which means I have a chance to
actually fix things via tests.