A quick note about how I'd *like* to do things.
Basically, if I could procedurally declare my API, then all of the boilerplate
could be easily generated from that. However, it needs to be 'static' in a C
header file, and to get that you'd have to generate it from some other source
or have a *lot* more voodoo in your C pre-processor code. Shame that C sucks so
much. Note that Pyrex gets this right, where doing all the boilerplate amounts
to adding 'api' to your definition. Aka
    cdef api object myfunc(int foo)
Is enough to generate all the boilerplate C headers, etc you would need.
It is unfortunate I can't really use that here, because I want a pure C class
to avoid the python garbage collector.