Minor tweak to Inventory.path2id
Name the parameter 'relpath' rather than 'name', as 'name' generally means
just the last component of the path.
Change the CHKInventory.path2id code to cache all of the intermediate
values, rather than just caching the final value.
Also change the lookup loop to use the cache if it is available.
Note that ideally we would actually check the cache in reverse order
first, and then build up whatever is missing afterwards...