Work on doing globbing, etc for all commands on Windows.
Implement a UnicodeShlex that takes the key bits from shlex.shlex.
Most notably, though, it works on a Unicode string, and not a collection
of input files. It also has a slightly different handling of \ because
those are frequently part of the path.

It also lets us know if a given section was quoted, so that we know if
we should skip glob expansion or not.

Then start expanding globs as appropriate.