Use the new KnownGraph.add_node() functionality.
This changes the file_id heads lookups to use a KnownGraph over all revision-ids
which should be a lot better than using a regular Graph instance.
Initial testing w/ xserver doesn't show much, since it seems to be purely
linear for quite a bit of x history.

Oddly enough the heads result seems to give a different number of file texts
which is worrying.