Add a helper _ModuleFeature.
It is *very* common for us to what to run a test or not based on
whether a module can be imported. So make it a common base class
rather than reimplementing it each time.

Also, when the module is available, you can get it from the
feature object, rather than doing your own import test.