Change the buffering algorithm for ProtocolThree.
Instead of flushing when we have had 100 write calls, flush when we have 1MB.
There are a lot of potential times when we could flush.
osutils.send_all() splits up the request into 64kB chunks. So we could
buffer to 64kB and then call flush. My concern there is for 'rounding'
errors. Where we end up with 65kB and then we have a nice big 64kB
send, and then a tiny 1kB send. Hopefully this will be a little bit better.

Another option which I might try, is to flush at >64kB, but round to a
64kB boundary, and keep any bytes that are past for the future.