More flexible handling of fixed word lists.
Now a command (like help) which accepts a known fixed set of non-option
arguments can simply set the bash variable fixedWords and have them
expanded.  All other commands fall back to default expansion unless the
current word starts with a '-'.

This avoids our need for the extglob shopt.  It also allows us to add more
such fixed wordlist commands in the future.  There might be a slight
performance impact as the case distinction is now executed even for the
completion of non-option arguments, but that shouldn't be noticable.