Provide --authors argument to log command.
In some situations it might be useful what authorship to display in a log. 
This change introduces a new --authors switch, with possible values 'all',
'first' and 'committer'.  Using this switch, user can override the choice
made by the log format.

On the implementation side in log.py, there is a new method
LogFormatter.authors which does all the author selection stuff.  Other
methods, in derived classes in particular, delegate to that method to do the
formatting.  This ensures that the override from the comand line applies to
all formats.